+++ src/GraphQL.SystemReactive/ExecutionResultExtensions.cs
