+++ src/GraphQL.SystemReactive/GraphQL.SystemReactive.csproj
