+++ src/GraphQL.SystemReactive/SubscriptionDocumentExecuter.cs
