+++ src/GraphQL.SystemReactive/SubscriptionExecutionStrategy.cs
