+++ src/GraphQL.Tests/Subscription/SubscriptionWithReflectionTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reactive.Linq;
 using System.Threading.Tasks;
+using GraphQL.Subscription;
 using Shouldly;
 using Xunit;
