+++ src/GraphQL.sln
@@ -119,7 +119,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "migrations", "migrations", 
 		..\docs2\site\docs\migrations\v0_8_0.md = ..\docs2\site\docs\migrations\v0_8_0.md
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.Subscriptions", "GraphQL.Subscriptions\GraphQL.Subscriptions.csproj", "{3AE97815-AEF3-445C-A617-A6863B46D96A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.SystemReactive", "GraphQL.SystemReactive\GraphQL.SystemReactive.csproj", "{B0E1FC8E-3112-4B9B-B61C-81F2A446379A}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Subscriptions", "Subscriptions", "{F9CEC116-8C0D-422D-823E-CC41230BE0E6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -171,10 +173,10 @@ Global
 		{FCE8E24C-29AA-4AED-93BA-08551D38823A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FCE8E24C-29AA-4AED-93BA-08551D38823A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FCE8E24C-29AA-4AED-93BA-08551D38823A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3AE97815-AEF3-445C-A617-A6863B46D96A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3AE97815-AEF3-445C-A617-A6863B46D96A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3AE97815-AEF3-445C-A617-A6863B46D96A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3AE97815-AEF3-445C-A617-A6863B46D96A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B0E1FC8E-3112-4B9B-B61C-81F2A446379A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B0E1FC8E-3112-4B9B-B61C-81F2A446379A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B0E1FC8E-3112-4B9B-B61C-81F2A446379A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B0E1FC8E-3112-4B9B-B61C-81F2A446379A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -192,6 +194,7 @@ Global
 		{4B4D88E2-B20E-45E0-94E5-D9DD19BDDC29} = {2A7FA3F9-B04C-463E-80EF-144433458FAD}
 		{FF7D1FB7-2288-4D87-989F-1E9D7BFB0AAF} = {2A7FA3F9-B04C-463E-80EF-144433458FAD}
 		{74E5F778-9FF7-42EE-A4D8-74509F79931F} = {EB92DFCF-A87E-4E8B-B068-8B396D00945A}
+		{B0E1FC8E-3112-4B9B-B61C-81F2A446379A} = {F9CEC116-8C0D-422D-823E-CC41230BE0E6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {61AAA7F9-9D46-4050-A3A5-F4358BBD95B8}
